feat(dashboard): add dashboard components and utilities
- Create new utility functions for time formatting, mode icons, and gradients
- Implement StatCard, ModeCard, and CompletedList components
- Refactor dashboard layout with new components and styling
- Add helper functions for mode details and display
- Improve type safety and code organization
