+++ src/components/dashboard/utils/formatTime.ts
@@ -1,0 +1,5 @@
+export const formatTime = (timeInSeconds: number): string => {
+    const minutes = Math.floor(timeInSeconds / 60)
+    const seconds = timeInSeconds % 60
+    return `${minutes}:${seconds.toString().padStart(2, "0")}`
+  }
